Changed int to long because sql uses Integer64bit and avoid future mistakes

+++ app/src/main/java/com/cr_d/passwordmanagerapp/application/interfaces/IPasswordRepository.kt
@@ -7,8 +7,8 @@ interface IPasswordRepository {
     fun findAll(): List<PasswordData>
     fun findByApplication(app: String): List<PasswordData>
     fun findByAccount(account: String): List<PasswordData>
-    fun findById(id: Int): PasswordData?
+    fun findById(id: Long): PasswordData?
     fun save(passwordData: PasswordEntity)
     fun update(passwordData: PasswordEntity)
-    fun delete(id: Int)
+    fun delete(id: Long)
 }
